re-add zarr
